+++ lib/communities/bloc/communities_bloc.dart
@@ -1,8 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:lemmy_api_client/v3.dart';
 
-import 'package:lemmy/lemmy.dart';
 import 'package:thunder/account/models/account.dart';
 import 'package:thunder/core/auth/helpers/fetch_account.dart';
 
@@ -15,17 +14,15 @@ class CommunitiesBloc extends Bloc<CommunitiesEvent, CommunitiesState> {
   CommunitiesBloc() : super(const CommunitiesState()) {
     on<ListCommunitiesEvent>((event, emit) async {
       Account? account = await fetchActiveProfileAccount();
-      Lemmy lemmy = LemmyClient.instance.lemmy;
+      LemmyApiV3 lemmy = LemmyClient.instance.lemmyApiV3;
 
-      ListCommunitiesResponse listCommunitiesResponse = await lemmy.listCommunities(
-        ListCommunities(
-          auth: account?.jwt,
-          page: state.page,
-          limit: 30,
-        ),
-      );
+      List<CommunityView> communities = await lemmy.run(ListCommunities(
+        auth: account?.jwt,
+        page: state.page,
+        limit: 30,
+      ));
 
-      return emit(state.copyWith(status: CommunitiesStatus.success, communities: listCommunitiesResponse.communities, page: state.page + 1));
+      return emit(state.copyWith(status: CommunitiesStatus.success, communities: communities, page: state.page + 1));
     });
   }
 }