+++ lib/shared/common_markdown_body.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:thunder/shared/image_preview.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 
 import 'package:thunder/account/bloc/account_bloc.dart';
@@ -19,9 +21,22 @@ class CommonMarkdownBody extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
+    bool openInExternalBrowser = false;
+
+    try {
+      context.read<ThunderBloc>().state.preferences?.getBool('setting_links_open_in_external_browser') ?? false;
+    } catch (e) {}
 
     return MarkdownBody(
       data: body,
+      imageBuilder: (uri, title, alt) {
+        return ImagePreview(
+          url: uri.toString(),
+          width: MediaQuery.of(context).size.width - 24,
+          isExpandable: true,
+          showFullHeightImages: true,
+        );
+      },
       selectable: isSelectableText,
       onTapLink: (text, url, title) {
         String? communityName = checkLemmyInstanceUrl(text);
@@ -44,7 +59,11 @@ class CommonMarkdownBody extends StatelessWidget {
             ),
           );
         } else if (url != null) {
-          Navigator.of(context).push(MaterialPageRoute(builder: (context) => WebView(url: url)));
+          if (openInExternalBrowser == true) {
+            launchUrl(Uri.parse(url), mode: LaunchMode.externalApplication);
+          } else {
+            Navigator.of(context).push(MaterialPageRoute(builder: (context) => WebView(url: url)));
+          }
         }
       },
       styleSheet: MarkdownStyleSheet.fromTheme(theme).copyWith(